refactor: Changed folder structure for redux features;
