+++ src/app/components/molecules/auth/LoginForm/LoginForm.tsx
@@ -6,7 +6,7 @@ import { AuthFormTemplate } from "@components/templates";
 import { IAuthFormProps } from "../IAuthFormProps";
 import { validateEmail, validateLength, validatePasswordStrength } from "@utils"
 import { useAppDispatch } from "@store";
-import { logIn } from "@slices";
+import { logIn } from "@features/auth";
 
 const LoginForm: React.FC<IAuthFormProps> = ({ toggleAuthMode }) => {
 	const [form, setField, resetForm, errors, validateAll] = useFormState({ email: '', password: '' }, { email: [validateEmail], password: [validateLength(), validatePasswordStrength]});
