+++ src/app/shared/features/auth/authSelectors.ts
@@ -1,4 +1,4 @@
-import { RootState } from "../store";
+import { RootState } from "@store";
 
 export const selectCurrentUser = (state: RootState) => state.auth.currentUser;
 export const selectAuthError = (state: RootState) => state.auth.error;
