+++ src/app/shared/features/auth/index.ts
@@ -1,0 +1,4 @@
+export * from "./authSelectors";
+export { default as authReducer } from "./authSlice";
+export * from "./authSlice";
+export * from "./authThunks";
