+++ src/app/shared/features/cart/cartSelectors.ts
@@ -1,4 +1,4 @@
-import { RootState } from "../store";
+import { RootState } from "@store";
 
 export const selectCartItems = (state: RootState) => state.cart.items;
 export const selectTotalItems = (state: RootState) => state.cart.items.reduce((sum, item) => sum + item.amount, 0);
