+++ src/app/shared/features/cart/index.ts
@@ -1,0 +1,3 @@
+export * from "./cartSelectors";
+export { default as cartReducer } from "./cartSlice";
+export * from "./cartSlice";
