+++ src/app/shared/features/fetch/fetchSelectors.ts
@@ -1,4 +1,4 @@
-import { emptyState } from "@slices";
 import { RootState } from "@store";
+import { emptyState } from "./fetchSlice";
 
 export const selectFetchResult = (key: string) => (state: RootState) => state.fetch[key] ?? emptyState;
