+++ src/app/shared/features/fetch/index.ts
@@ -1,0 +1,4 @@
+export * from "./fetchSelectors";
+export { default as fetchReducer} from "./fetchSlice";
+export * from "./fetchSlice";
+export * from "./fetchThunks";
