+++ src/app/shared/hooks/index.ts
@@ -1,2 +1,1 @@
-export { useFetch } from "./useFetch";
 export { useFormState } from "./useFormState";
