+++ src/app/shared/interfaces/IUser.ts
@@ -1,0 +1,5 @@
+export interface IUser {
+  email: string | null;
+  displayName: string | null;
+  photoURL: string | null;
+}
