+++ src/app/shared/interfaces/index.ts
@@ -1,2 +1,3 @@
 export * from "./IProduct";
 export * from "./ICartItem";
+export * from "./IUser";
