--- src/app/shared/selectors/index.ts
@@ -1,3 +1,0 @@
-export * from "./cartSelectors";
-export * from "./authSelectors";
-export * from "./fetchSelectors";
