--- src/app/shared/slices/index.ts
@@ -1,6 +1,0 @@
-export { default as cartReducer } from "./cartSlice";
-export { default as authReducer } from "./authSlice";
-export { default as fetchReducer} from "./fetchSlice";
-export * from "./cartSlice";
-export * from "./authSlice";
-export * from "./fetchSlice";
