+++ src/app/shared/store/hooks.ts
@@ -1,0 +1,5 @@
+import { useDispatch, TypedUseSelectorHook, useSelector } from "react-redux";
+import { AppDispatch, RootState } from "./store";
+
+export const useAppDispatch = () => useDispatch<AppDispatch>();
+export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
