+++ src/app/shared/store/store.ts
@@ -1,6 +1,7 @@
+import { authReducer } from "@features/auth";
+import { cartReducer } from "@features/cart";
+import { fetchReducer } from "@features/fetch";
 import { configureStore } from "@reduxjs/toolkit";
-import { authReducer, cartReducer, fetchReducer } from "@slices";
-import { useDispatch, TypedUseSelectorHook, useSelector } from "react-redux";
 
 export const store = configureStore({
   reducer: {
@@ -12,5 +13,3 @@ export const store = configureStore({
 
 export type RootState = ReturnType<typeof store.getState>;
 export type AppDispatch = typeof store.dispatch;
-export const useAppDispatch = () => useDispatch<AppDispatch>();
-export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
