+++ src/app/shared/utils/firebaseHelpers.ts
@@ -1,0 +1,8 @@
+import { IUser } from "@interfaces";
+import { User } from "firebase/auth";
+
+export const mapFirebaseUserToUser = (user: User): IUser => ({
+  email: user.email,
+  displayName: user.displayName,
+  photoURL: user.photoURL,
+});
